Move model_path fixture to top level conftest.py
