+++ tests/catboost/test_catboost_model_export.py
@@ -92,11 +92,6 @@ def get_reg_model_signature():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def custom_env(tmp_path):
     conda_env_path = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env_path, additional_pip_deps=["catboost", "pytest"])
