+++ tests/fastai/test_fastai_model_export.py
@@ -50,11 +50,6 @@ def fastai_model():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def fastai_custom_env(tmp_path):
     conda_env = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["fastai", "pytest"])
