+++ tests/h2o/test_h2o_model_export.py
@@ -77,11 +77,6 @@ def h2o_iris_model_signature():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def h2o_custom_env(tmp_path):
     conda_env = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["h2o", "pytest"])
