+++ tests/paddle/test_paddle_model_export.py
@@ -107,11 +107,6 @@ def pd_model_signature():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def pd_custom_env(tmp_path):
     conda_env = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["paddle", "pytest"])
