+++ tests/prophet/test_prophet_model_export.py
@@ -124,11 +124,6 @@ def prophet_model():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return tmp_path.joinpath("model")
-
-
-@pytest.fixture
 def prophet_custom_env(tmp_path):
     conda_env = tmp_path.joinpath("conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["prophet"])
