+++ tests/tensorflow/test_keras_model_export.py
@@ -171,11 +171,6 @@ def custom_predicted(custom_model, data):
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def keras_custom_env(tmp_path):
     conda_env = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["keras", "tensorflow", "pytest"])
