+++ tests/xgboost/test_xgboost_model_export.py
@@ -81,11 +81,6 @@ def xgb_sklearn_model():
 
 
 @pytest.fixture
-def model_path(tmp_path):
-    return os.path.join(tmp_path, "model")
-
-
-@pytest.fixture
 def xgb_custom_env(tmp_path):
     conda_env = os.path.join(tmp_path, "conda_env.yml")
     _mlflow_conda_env(conda_env, additional_pip_deps=["xgboost", "pytest"])
